Move all exceptions on Exception namespace
