--- Driver/DriverUnreachableResourceException.php
@@ -1,20 +1,0 @@
-<?php
-
-/**
- * Rss/Atom Bundle for Symfony 2.
- *
- *
- * @license http://opensource.org/licenses/lgpl-3.0.html LGPL
- * @copyright (c) 2013, Alexandre Debril
- */
-
-namespace Debril\RssAtomBundle\Driver;
-
-use Debril\RssAtomBundle\Exception\RssAtomException;
-
-/**
- * Class DriverUnreachableResourceException.
- */
-class DriverUnreachableResourceException extends RssAtomException
-{
-}
