+++ Exception/FeedException/FeedCannotBeReadException.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * Rss/Atom Bundle for Symfony 2.
+ *
+ *
+ * @license http://opensource.org/licenses/lgpl-3.0.html LGPL
+ * @copyright (c) 2013, Alexandre Debril
+ */
+
+namespace Debril\RssAtomBundle\Exception\FeedException;
+
+use Debril\RssAtomBundle\Exception\FeedException;
+
+/**
+ * Class FeedCannotBeReadException.
+ */
+class FeedCannotBeReadException extends FeedException
+{
+}
