+++ Provider/MockProvider.php
@@ -13,7 +13,7 @@ namespace Debril\RssAtomBundle\Provider;
 
 use Debril\RssAtomBundle\Protocol\Parser\FeedContent;
 use Debril\RssAtomBundle\Protocol\Parser\Item;
-use Debril\RssAtomBundle\Exception\FeedNotFoundException;
+use Debril\RssAtomBundle\Exception\FeedException\FeedNotFoundException;
 
 class MockProvider implements FeedContentProviderInterface
 {
