+++ Tests/Driver/HttpCurlDriverTest.php
@@ -2,6 +2,8 @@
 
 namespace Debril\RssAtomBundle\Driver;
 
+use Debril\RssAtomBundle\Exception\DriverUnreachableResourceException;
+
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.0 on 2013-02-06 at 23:48:53.
  */
@@ -54,7 +56,7 @@ class HttpCurlDriverTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers Debril\RssAtomBundle\Driver\HttpCurlDriver::getResponse
-     * @expectedException Debril\RssAtomBundle\Driver\DriverUnreachableResourceException
+     * @expectedException \Debril\RssAtomBundle\Exception\DriverUnreachableResourceException
      */
     public function testGetResponseException()
     {
